+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 - calculateEVMTokenAddress function now requires tokenUtils address as deployer and crossChainLayer address as constructor params
 
+- Rename shardedId -> shardsKey
+
 ## [0.4.2] - 2025-02-05
 
 ## Added